requirements.txt
tensorflow-cpu==2.17.1
numpy==1.26.4
Pillow==10.2.0
